feat: add generic resource-store crate with TAO-inspired store abstractions
Introduce `crates/resource-store`, a reusable crate providing the core
abstractions for a graph-based resource store:

- `Label` trait for type-safe resource type discriminants
- `Object<L>` / `Association<L>` generic graph node and edge types
- `ObjectStore<L>` / `AssociationStore<L>` async CRUD + graph traits
- `SecretManager` trait for encrypted sensitive field storage
- `FieldRole` / `ResourceTypeDescriptor` / `ResourceRegistry` for
  proto-driven field role metadata (data, identifier, sensitive, managed)
- `ResourceName` and `ResourceRef` moved from `crates/common` to
  `resource-store` (common re-exports for backward compatibility)

The crate is generic over `L: Label` (one type parameter) so downstream
projects can generate their own label enums from proto definitions while
sharing the store infrastructure.
