+++ crates/resource-store/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "unitycatalog-resource-store"
+edition.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+license.workspace = true
+repository.workspace = true
+readme.workspace = true
+version.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+async-trait = { workspace = true }
+bytes = { workspace = true }
+chrono = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+thiserror = { workspace = true }
+uuid = { workspace = true }
+
+# convenience for downstream crates that want to use certain types with sqlx
+sqlx = { workspace = true, optional = true, features = [
+  "uuid",
+  "macros",
+  "chrono",
+  "json",
+] }
+
+[features]
+default = []
+sqlx = ["dep:sqlx"]
